_save 5 & fix ub

+++ c/audio/audio.c
@@ -3,14 +3,14 @@ void smpt_adoMgen(float *Pdata, int32_t Iframe)
 {
 	const float Ffreq = 440.0F;
 	const float Fvolume = 0.2F;
-	const float Fstep = 2.0F * M_PI * Ffreq / SMPT_ADOfSAMPLE_RATE;
+	const float Fstep = 2.0F * (float)M_PI * Ffreq / SMPT_ADOfSAMPLE_RATE;
 	for (int32_t I0 = 0; I0 < Iframe; I0++)
 	{
 			float Fs = sinf(Fphase) * Fvolume;
 			Fphase += Fstep;
 
-			if (Fphase > 2.0F * M_PI)
-				Fphase -= 2.0F * M_PI;
+			if (Fphase > 2.0F * (float)M_PI)
+				Fphase -= 2.0F * (float)M_PI;
 
 			//.i left
 			Pdata[I0 * SMPT_ADOuCHANNEL_COUNT] = Fs;
